+++ src/main/java/com/acme/stonks/resource/AccountTermDepositResource.java
@@ -5,8 +5,8 @@ import java.util.Date;
 public class AccountTermDepositResource {
 
     private Long id;
-    private float capital;
-    private float tea;
+    private double capital;
+    private double tea;
     private float minTea;
     private Date dateStart;
     private Date dateEnd;
@@ -19,16 +19,16 @@ public class AccountTermDepositResource {
     public void setId(Long id) {
         this.id = id;
     }
-    public float getCapital() {
+    public double getCapital() {
         return capital;
     }
-    public void setCapital(float capital) {
+    public void setCapital(double capital) {
         this.capital = capital;
     }
-    public float getTea() {
+    public double getTea() {
         return tea;
     }
-    public void setTea(float tea) {
+    public void setTea(double tea) {
         this.tea = tea;
     }
     public float getMinTea() {
